:muscle: Move commands into `command` module
