+++ denops/gin/command/bare/command.ts
