+++ denops/gin/command/bare/main.ts
