+++ denops/gin/command/branch/action.ts
@@ -3,7 +3,7 @@ import * as fn from "https://deno.land/x/denops_std@v4.0.0/function/mod.ts";
 import * as batch from "https://deno.land/x/denops_std@v4.0.0/batch/mod.ts";
 import * as helper from "https://deno.land/x/denops_std@v4.0.0/helper/mod.ts";
 import { define, Range } from "../../core/action/action.ts";
-import { command as commandBare } from "../../core/bare/command.ts";
+import { command as commandBare } from "../../command/bare/command.ts";
 import { Branch, parse as parseBranch } from "./parser.ts";
 
 export async function init(denops: Denops, bufnr: number): Promise<void> {
