+++ denops/gin/command/buffer/edit.ts
