+++ denops/gin/command/buffer/read.ts
