+++ denops/gin/command/diff/command.ts
