+++ denops/gin/command/diff/commitish.ts
