+++ denops/gin/command/diff/commitish_test.ts
