+++ denops/gin/command/diff/jump.ts
