+++ denops/gin/command/diff/jump_test.ts
