+++ denops/gin/command/diff/main.ts
