+++ denops/gin/command/edit/command.ts
