+++ denops/gin/command/edit/main.ts
