+++ denops/gin/command/edit/read.ts
