+++ denops/gin/command/edit/util.ts
