+++ denops/gin/command/log/command.ts
