+++ denops/gin/command/log/edit.ts
@@ -1,4 +1,5 @@
 import type { Denops } from "https://deno.land/x/denops_std@v4.0.0/mod.ts";
+import { unnullish } from "https://deno.land/x/unnullish@v1.0.0/mod.ts";
 import * as batch from "https://deno.land/x/denops_std@v4.0.0/batch/mod.ts";
 import * as buffer from "https://deno.land/x/denops_std@v4.0.0/buffer/mod.ts";
 import * as option from "https://deno.land/x/denops_std@v4.0.0/option/mod.ts";
@@ -9,10 +10,10 @@ import {
   formatFlags,
   parseOpts,
 } from "https://deno.land/x/denops_std@v4.0.0/argument/mod.ts";
-import { bind } from "../../core/bare/command.ts";
-import { exec as execBuffer } from "../../core/buffer/edit.ts";
+import { bind } from "../../command/bare/command.ts";
+import { exec as execBuffer } from "../../command/buffer/edit.ts";
 import { init as initActionCore } from "../../core/action/action.ts";
-import { init as initActionBranch } from "./action.ts";
+import { init as initActionLog } from "./action.ts";
 
 export async function edit(
   denops: Denops,
@@ -25,7 +26,7 @@ export async function edit(
   await exec(denops, bufnr, {
     worktree: expr,
     flags: params,
-    patterns: fragment?.replace(/\$$/, "").split(" ").filter((v) => v),
+    args: unnullish(fragment, (v) => JSON.parse(v.replace(/\$$/, ""))),
     encoding: opts.enc ?? opts.encoding,
     fileformat: opts.ff ?? opts.fileformat,
   });
@@ -34,7 +35,7 @@ export async function edit(
 export type ExecOptions = {
   worktree?: string;
   flags?: Flags;
-  patterns?: string[];
+  args?: string[];
   encoding?: string;
   fileformat?: string;
 };
@@ -45,12 +46,10 @@ export async function exec(
   options: ExecOptions,
 ): Promise<void> {
   const args = [
-    "branch",
-    "--list",
-    "-vv",
+    "log",
+    "--color=always",
     ...formatFlags(options.flags ?? {}),
-    "--",
-    ...(options.patterns ?? []),
+    ...(options.args ?? []),
   ];
   await execBuffer(denops, bufnr, args, {
     worktree: options.worktree,
@@ -61,8 +60,8 @@ export async function exec(
     await batch.batch(denops, async (denops) => {
       await bind(denops, bufnr);
       await initActionCore(denops, bufnr);
-      await initActionBranch(denops, bufnr);
-      await option.filetype.setLocal(denops, "gin-branch");
+      await initActionLog(denops, bufnr);
+      await option.filetype.setLocal(denops, "gin-log");
     });
   });
 }
