+++ denops/gin/command/log/main.ts
