+++ denops/gin/command/log/parser.ts
