+++ denops/gin/command/log/parser_test.ts
