+++ denops/gin/command/patch/command.ts
