+++ denops/gin/command/patch/main.ts
