+++ denops/gin/command/status/command.ts
