+++ denops/gin/command/status/edit.ts
@@ -1,5 +1,7 @@
 import type { Denops } from "https://deno.land/x/denops_std@v4.0.0/mod.ts";
 import { unnullish } from "https://deno.land/x/unnullish@v1.0.0/mod.ts";
+import * as autocmd from "https://deno.land/x/denops_std@v4.0.0/autocmd/mod.ts";
+import * as fn from "https://deno.land/x/denops_std@v4.0.0/function/mod.ts";
 import * as batch from "https://deno.land/x/denops_std@v4.0.0/batch/mod.ts";
 import * as buffer from "https://deno.land/x/denops_std@v4.0.0/buffer/mod.ts";
 import * as option from "https://deno.land/x/denops_std@v4.0.0/option/mod.ts";
@@ -10,10 +12,10 @@ import {
   formatFlags,
   parseOpts,
 } from "https://deno.land/x/denops_std@v4.0.0/argument/mod.ts";
-import { bind } from "../../core/bare/command.ts";
-import { exec as execBuffer } from "../../core/buffer/edit.ts";
+import { bind } from "../../command/bare/command.ts";
+import { exec as execBuffer } from "../../command/buffer/edit.ts";
 import { init as initActionCore } from "../../core/action/action.ts";
-import { init as initActionLog } from "./action.ts";
+import { init as initActionStatus } from "./action.ts";
 
 export async function edit(
   denops: Denops,
@@ -26,7 +28,7 @@ export async function edit(
   await exec(denops, bufnr, {
     worktree: expr,
     flags: params,
-    args: unnullish(fragment, (v) => JSON.parse(v.replace(/\$$/, ""))),
+    pathspecs: unnullish(fragment, (v) => JSON.parse(v.replace(/\$$/, ""))),
     encoding: opts.enc ?? opts.encoding,
     fileformat: opts.ff ?? opts.fileformat,
   });
@@ -35,7 +37,7 @@ export async function edit(
 export type ExecOptions = {
   worktree?: string;
   flags?: Flags;
-  args?: string[];
+  pathspecs?: string[];
   encoding?: string;
   fileformat?: string;
 };
@@ -46,10 +48,13 @@ export async function exec(
   options: ExecOptions,
 ): Promise<void> {
   const args = [
-    "log",
-    "--color=always",
+    "status",
+    "--short",
+    "--branch",
+    "--ahead-behind",
     ...formatFlags(options.flags ?? {}),
-    ...(options.args ?? []),
+    "--",
+    ...(options.pathspecs ?? []),
   ];
   await execBuffer(denops, bufnr, args, {
     worktree: options.worktree,
@@ -57,11 +62,28 @@ export async function exec(
     fileformat: options.fileformat,
   });
   await buffer.ensure(denops, bufnr, async () => {
+    await buffer.modifiable(denops, bufnr, async () => {
+      const saved = await fn.winsaveview(denops);
+      await denops.cmd("silent! 2,$sort /.. /)");
+      await fn.winrestview(denops, saved);
+    });
     await batch.batch(denops, async (denops) => {
       await bind(denops, bufnr);
       await initActionCore(denops, bufnr);
-      await initActionLog(denops, bufnr);
-      await option.filetype.setLocal(denops, "gin-log");
+      await initActionStatus(denops, bufnr);
+      await option.filetype.setLocal(denops, "gin-status");
+      await autocmd.group(
+        denops,
+        `gin_command_status_command_read_${bufnr}`,
+        (helper) => {
+          helper.remove();
+          helper.define(
+            ["BufWritePost", "FileWritePost"],
+            "*",
+            `call gin#util#reload(${bufnr})`,
+          );
+        },
+      );
     });
   });
 }
