+++ denops/gin/command/status/main.ts
