+++ denops/gin/command/status/parser.ts
