+++ denops/gin/command/status/parser_test.ts
