+++ denops/gin/main.ts
@@ -1,18 +1,18 @@
 import type { Denops } from "https://deno.land/x/denops_std@v4.0.0/mod.ts";
 
 import { main as mainAction } from "./core/action/main.ts";
-import { main as mainBare } from "./core/bare/main.ts";
-import { main as mainBuffer } from "./core/buffer/main.ts";
+import { main as mainBare } from "./command/bare/main.ts";
+import { main as mainBuffer } from "./command/buffer/main.ts";
 import { main as mainProxy } from "./proxy/main.ts";
 import { main as mainUtil } from "./util/main.ts";
 
-import { main as mainBranch } from "./feat/branch/main.ts";
-import { main as mainChaperon } from "./feat/chaperon/main.ts";
-import { main as mainDiff } from "./feat/diff/main.ts";
-import { main as mainEdit } from "./feat/edit/main.ts";
-import { main as mainLog } from "./feat/log/main.ts";
-import { main as mainPatch } from "./feat/patch/main.ts";
-import { main as mainStatus } from "./feat/status/main.ts";
+import { main as mainBranch } from "./command/branch/main.ts";
+import { main as mainChaperon } from "./command/chaperon/main.ts";
+import { main as mainDiff } from "./command/diff/main.ts";
+import { main as mainEdit } from "./command/edit/main.ts";
+import { main as mainLog } from "./command/log/main.ts";
+import { main as mainPatch } from "./command/patch/main.ts";
+import { main as mainStatus } from "./command/status/main.ts";
 
 import { main as mainComponentBranch } from "./component/branch.ts";
 import { main as mainComponentTraffic } from "./component/traffic.ts";
